test: move the thinking-block-binding suite into `tests/models/anthropic/`
The binding tests, their helpers and their cassettes were appended to the
already-large per-provider `test_anthropic.py`. The testing guidelines
steer a self-contained feature toward a feature-centric file, and
`tests/models/anthropic/` already holds siblings like
`test_code_execution_files.py`, so the suite and its fixtures now live in
`test_thinking_block_binding.py` with their cassettes alongside.
