+++ tests/models/anthropic/cassettes/test_thinking_block_binding/test_anthropic_fable_5_1_drops_a_stale_thinking_block.yaml
