+++ tests/models/anthropic/cassettes/test_thinking_block_binding/test_anthropic_fable_5_1_replays_a_stale_thinking_block_on_a_legacy_account.yaml
